Implement some big chunk of work
* Expose commit ids via KV methods. The mutation and query methods on KV have
overloads to return results wrapped in `Ordered` objects. `Ordered` contains a
result along with the commit index on which that result is produced.
* Enable local queries. The query methods on KV have overloads to perform a
local query with preserving consistency of the observed state. When a commit
index is provided to these methods, the attempted query operation is executed on
the connected AfloatDB server only if the server's Raft state is sufficiently up
to date.
* Eliminate unnecessary object creations in InvocationService for local Raft
invocations.

Signed-off-by: Ensar Basri Kahveci <ebkahveci@gmail.com>