Add timeout parameter to RaftNode#query() method
This commit implements a new RaftNode#query() override which contains a timeout
parameter. When a monotonic reads query is issued with
QueuePolicy#EVENTUAL_CONSISTENCY, a minimum commit index and a timeout,
RaftNode waits for the given timeout duration until the local commit index
becomes greater than or equal to the given minimum commit index.