Fix a linearizable query commit index bug
This commit fixes a rare bug that can occur during membership changes.
The fix is validated in the new test case added to LinearizableQueryTest.java.
When a membership change is appended to the leader but not yet committed,
if the current commit index is X and the last log index is Y where Y > X,
when a linearizable query is executed, Y is passed to
StateMachine.runOperation() instead of X. This does not violate any consistency
property since the query is executed on the latest commit state of
the StateMachine, however the incorrect commit index passed to StateMachine
might be used for secondary purposes, such as logging, etc.

Secondarily, this commit fixes another small bug where a leader Raft node does
not gracefully shutdown its components automatically which required
a RaftNode#terminate() call after the membership change is committed.