Fixed an issue where the old leader's log could not be synced with the new one due to the absence of a previous LogEntry
If an old leader's log contains a snapshot and only unreplicated entries, it can't be synchronized with a new leader's log.
This happens because the new leader can't verify the previous LogEntry of the old leader since this entry is already snapshotted.
As a fix, we should add a check if `snapshotIndex` is large or equal to `previousLogIndex`, and use snapshot term in this case.